[hbase] switch to using hbase major version numbers instead of minor version numbers.
