+++ asynchbase/pom.xml
@@ -97,14 +97,14 @@ LICENSE file.
     <dependency>
       <groupId>org.apache.hbase</groupId>
       <artifactId>hbase-shaded-testing-util</artifactId>
-      <version>${hbase14.version}</version>
+      <version>${hbase1.version}</version>
       <scope>test</scope>
     </dependency>
     
     <dependency>
       <groupId>org.apache.hbase</groupId>
       <artifactId>hbase-shaded-client</artifactId>
-      <version>${hbase14.version}</version>
+      <version>${hbase1.version}</version>
       <scope>test</scope>
     </dependency>
