+++ bin/bindings.properties
@@ -48,8 +48,8 @@ foundationdb:site.ycsb.db.foundationdb.FoundationDBClient
 geode:site.ycsb.db.GeodeClient
 googlebigtable:site.ycsb.db.GoogleBigtableClient
 googledatastore:site.ycsb.db.GoogleDatastoreClient
-hbase14:site.ycsb.db.hbase14.HBaseClient14
-hbase22:site.ycsb.db.hbase22.HBaseClient22
+hbase1:site.ycsb.db.hbase1.HBaseClient1
+hbase2:site.ycsb.db.hbase2.HBaseClient2
 hypertable:site.ycsb.db.HypertableClient
 ignite:site.ycsb.db.ignite.IgniteClient
 ignite-sql:site.ycsb.db.ignite.IgniteSqlClient
