+++ hbase2/src/test/resources/hbase-site.xml
